requirements.txt
numpy>=1.23
matplotlib>=3.7
PyYAML>=6.0
pandas>=2.0
pytest>=7.0

